Move TGGMutantRuleUtil to a separate file Pick a random mutant Save mutant actions to a cache
